+++ src/runtime/components/media/Media.vue
@@ -29,7 +29,9 @@
         :folder-type="folderType"
         :response-received="responseReceived"
         :request-pre-sent="requestPreSent"
+        :forward-request="forwardRequest"
         :alter-error-received="alterErrorReceived"
+        :accepted-file-types="acceptedFileTypes"
         @updateMediaComponent="onMediaComponentUpdate"
         @onMediaSelected="(media) => $emit('getSelectedMedia', media)"
     />
@@ -143,6 +145,14 @@ const props = defineProps({
   requestPreSent: {
     type: Function,
     default: () => {}
+  },
+  forwardRequest: {
+    type: Function,
+    default: null
+  },
+  acceptedFileTypes: {
+    type: String,
+    default: ''
   }
 })